+++ perl/FSM/Composition/Top.pm
@@ -11,21 +11,37 @@ use FSM::Composition::TopSymbols;
 sub new ($class, %args) {
     return bless {
         name => $args{name},
-        instances => $args{instances} || [],
-        ports_blocks => $args{ports_blocks} || [],
-        toplinks => $args{toplinks} || [],
-        package_imports => $args{package_imports} || [],
+        instances => _clone($args{instances} || []),
+        ports_blocks => _clone($args{ports_blocks} || []),
+        toplinks => _clone($args{toplinks} || []),
+        package_imports => _clone($args{package_imports} || []),
         top_symbols => $args{top_symbols} || FSM::Composition::TopSymbols->new(),
-        raw_ast => $args{raw_ast},
+        raw_ast => _clone($args{raw_ast}),
     }, $class;
 }
 
 sub name ($self) { return $self->{name} }
-sub instances ($self) { return $self->{instances} }
-sub ports_blocks ($self) { return $self->{ports_blocks} }
-sub toplinks ($self) { return $self->{toplinks} }
-sub package_imports ($self) { return $self->{package_imports} }
+sub instances ($self) { return _clone($self->{instances}) }
+sub ports_blocks ($self) { return _clone($self->{ports_blocks}) }
+sub toplinks ($self) { return _clone($self->{toplinks}) }
+sub package_imports ($self) { return _clone($self->{package_imports}) }
 sub top_symbols ($self) { return $self->{top_symbols} }
-sub raw_ast ($self) { return $self->{raw_ast} }
+sub raw_ast ($self) { return _clone($self->{raw_ast}) }
+
+sub _clone ($value) {
+    return undef unless defined $value;
+
+    if (ref($value) eq 'HASH') {
+        return {
+            map { $_ => _clone($value->{$_}) } sort keys %$value
+        };
+    }
+
+    if (ref($value) eq 'ARRAY') {
+        return [ map { _clone($_) } @$value ];
+    }
+
+    return $value;
+}
 
 1;